requirements.txt
annotated-types==0.7.0
anyio==4.6.2
bcrypt==4.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
passlib==1.7.4
pydantic==2.9.2
pydantic-core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-http-client==3.3.7
PyYAML==6.0.2
sendgrid==6.11.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.41.2
typing-extensions==4.12.2
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
